Make a proper test - sadly, it segfaults.
